+++ packages/typespec-autorest/test/azure-core-operations.test.ts
@@ -1,5 +1,5 @@
 import { deepStrictEqual } from "assert";
-import { describe, it } from "vitest";
+import { it } from "vitest";
 import { compileOpenAPI } from "./test-host.js";
 
 const wrapperCode = `
@@ -33,44 +33,11 @@ SupportsClientRequestId;
 alias Operations = Azure.Core.ResourceOperations<ServiceTraits>;
 `;
 
-describe("typespec-autorest: Azure.Core.ResourceOperations", () => {
-  it("ensure properties with 'create' visibility are included in the ResourceCreateOrUpdate body", async () => {
-    const result = await compileOpenAPI(
-      `
-        ${wrapperCode}
-    
-        
-        @resource("widgets")
-        model Widget {
-          @key("widgetName")
-          
-          @visibility(Lifecycle.Read)
-          name: string;
-        
-          
-          @visibility(Lifecycle.Create, Lifecycle.Read)
-          modality: string;
-        
-          
-          color: string;
-        }
-        
-        
-        @test
-        #suppress "@typespec/http/deprecated-implicit-optionality" "For test"
-        op createOrUpdateWidget is Operations.ResourceCreateOrUpdate<Widget>;
-      `,
-      { preset: "azure" },
-    );
-    const propKeys = Object.keys(result.definitions?.["WidgetCreateOrUpdate"].properties!);
-    deepStrictEqual(propKeys, ["modality", "color"]);
-  });
-
-  it("ensure properties with 'create' visibility are included in the LongRunningResourceCreateOrUpdate body", async () => {
-    const result = await compileOpenAPI(
-      `
+it("ensure properties with 'create' visibility are included in the ResourceCreateOrUpdate body", async () => {
+  const result = await compileOpenAPI(
+    `
       ${wrapperCode}
-
+  
       
       @resource("widgets")
       model Widget {
@@ -90,57 +57,88 @@ describe("typespec-autorest: Azure.Core.ResourceOperations", () => {
       
       @test
       #suppress "@typespec/http/deprecated-implicit-optionality" "For test"
-      op createOrUpdateWidget is Operations.LongRunningResourceCreateOrUpdate<Widget>;
+      op createOrUpdateWidget is Operations.ResourceCreateOrUpdate<Widget>;
     `,
-      { preset: "azure" },
-    );
+    { preset: "azure" },
+  );
+  const propKeys = Object.keys(result.definitions?.["WidgetCreateOrUpdate"].properties!);
+  deepStrictEqual(propKeys, ["modality", "color"]);
+});
 
-    const propKeys = Object.keys(result.definitions!["WidgetCreateOrUpdate"].properties!);
-    deepStrictEqual(propKeys, ["modality", "color"]);
-  });
+it("ensure properties with 'create' visibility are included in the LongRunningResourceCreateOrUpdate body", async () => {
+  const result = await compileOpenAPI(
+    `
+    ${wrapperCode}
 
-  it("ensure ConditionalRequestHeaders does not appear on Action or List operations", async () => {
-    function checkParams(params: any, path: string) {
-      params.forEach((param: { $ref?: string }) => {
-        if (param.$ref) {
-          const hasConditionalRequestHeaders = param.$ref.indexOf("ConditionalRequestHeaders") > -1;
-          if (hasConditionalRequestHeaders) {
-            throw new Error(`ConditionalRequestHeaders should not appear in ${path}`);
-          }
+    
+    @resource("widgets")
+    model Widget {
+      @key("widgetName")
+      
+      @visibility(Lifecycle.Read)
+      name: string;
+    
+      
+      @visibility(Lifecycle.Create, Lifecycle.Read)
+      modality: string;
+    
+      
+      color: string;
+    }
+    
+    
+    @test
+    #suppress "@typespec/http/deprecated-implicit-optionality" "For test"
+    op createOrUpdateWidget is Operations.LongRunningResourceCreateOrUpdate<Widget>;
+  `,
+    { preset: "azure" },
+  );
+
+  const propKeys = Object.keys(result.definitions!["WidgetCreateOrUpdate"].properties!);
+  deepStrictEqual(propKeys, ["modality", "color"]);
+});
+
+it("ensure ConditionalRequestHeaders does not appear on Action or List operations", async () => {
+  function checkParams(params: any, path: string) {
+    params.forEach((param: { $ref?: string }) => {
+      if (param.$ref) {
+        const hasConditionalRequestHeaders = param.$ref.indexOf("ConditionalRequestHeaders") > -1;
+        if (hasConditionalRequestHeaders) {
+          throw new Error(`ConditionalRequestHeaders should not appear in ${path}`);
         }
-      });
+      }
+    });
+  }
+
+  const result = await compileOpenAPI(
+    `
+    ${wrapperCode}
+
+    
+    @resource("widgets")
+    model Widget {
+      @key("widgetName")
+      
+      name: string;
     }
 
-    const result = await compileOpenAPI(
-      `
-      ${wrapperCode}
+    
+    @test
+    op listWidgets is Operations.ResourceList<
+      Widget,
+      ListQueryParametersTrait<StandardListQueryParameters & SelectQueryParameter>
+    >;
 
-      
-      @resource("widgets")
-      model Widget {
-        @key("widgetName")
-        
-        name: string;
-      }
+    
+    @test
+    op actionWidget is Operations.ResourceAction<Widget, {}, {}>;
+  `,
+    { preset: "azure" },
+  );
 
-      
-      @test
-      op listWidgets is Operations.ResourceList<
-        Widget,
-        ListQueryParametersTrait<StandardListQueryParameters & SelectQueryParameter>
-      >;
+  let params = result.paths["/widgets/{widgetName}:actionWidget"].post!.parameters;
+  checkParams(params, "/widgets/{widgetName}:actionWidget");
 
-      
-      @test
-      op actionWidget is Operations.ResourceAction<Widget, {}, {}>;
-    `,
-      { preset: "azure" },
-    );
-
-    let params = result.paths["/widgets/{widgetName}:actionWidget"].post!.parameters;
-    checkParams(params, "/widgets/{widgetName}:actionWidget");
-
-    params = result.paths["/widgets"].get!.parameters;
-    checkParams(params, "/widgets");
-  });
+  params = result.paths["/widgets"].get!.parameters;
+  checkParams(params, "/widgets");
 });