+++ packages/typespec-ts/src/transform/transform-telemetry-info.ts
@@ -1,0 +1,33 @@
+import {
+  getHttpOperationWithCache,
+  SdkClient,
+  SdkContext,
+} from "@azure-tools/typespec-client-generator-core";
+import { TelemetryInfo } from "../interfaces.js";
+import { listOperationsUnderClient } from "../utils/client-utils.js";
+import { getCustomRequestHeaderNameForOperation } from "../utils/operation-util.js";
+
+export function transformTelemetryInfo(
+  client: SdkClient,
+  dpgContext: SdkContext,
+): TelemetryInfo | undefined {
+  const customRequestIdHeaderName = getCustomRequestHeaderNameForClient(dpgContext, client);
+  if (customRequestIdHeaderName) {
+    return {
+      customRequestIdHeaderName,
+    };
+  }
+  return undefined;
+}
+
+function getCustomRequestHeaderNameForClient(dpgContext: SdkContext, client: SdkClient) {
+  for (const op of listOperationsUnderClient(client)) {
+    const headerName = getCustomRequestHeaderNameForOperation(
+      getHttpOperationWithCache(dpgContext, op),
+    );
+    if (headerName !== undefined) {
+      return headerName;
+    }
+  }
+  return undefined;
+}